script.js
A new function designed that should run whenever the reset button is clicked. It will do the following:

Check in the globalUniverse array if there is at least one verse from any of the 18 chapters. If not then add the full set of verses belonging to that chapter into the global universe array and even show an alert message informing the user about the chapter whose verses had got finished were added.

This function is called manually in the beginning in the resetRound() function.